feat(wip): add character property to control injury receiving
Decouples the `IsAffectedByInjuries` property from being used for both receiving injuries (in actor.nut onDamageReceived) and for applying the effects of already received injuries (in vanilla injuries' onUpdate functions return early if IsAffectedByInjuries is false thereby not applying their effect).

With this change the CanReceiveTemporaryInjuries property now controls receiving injuries, whereas the IsAffectedByInjuries and IsAffectedByFreshInjuries properties are confined to trigger the effects of received injuries.