fix: missing World keys in AfterHooks automatic hireable bg array setup (#24)
We temporarily instantiate and populate all the keys in ::World that are created by vanilla in world_state.onInit. We don't add a try/catch block because it fails to catch any error as the error occurs inside `create` so the try/catch doesn't stop it and throws the errors with stack trace into the log.