fix: preview skill potentially becoming null while previewing (#45)
This can happen due to mods which allow swapping items via a hotkey during a preview resulting in the previewed skill being removed from the actor. Therefore, now we store a WeakTableRef instead of a strong ref to the skill and only return true for `isPreviewing()` when either previewed skill or movement is not null.