fix(vanilla): calling onEntityEntersFirstSlot on entity pushed back
(#43)

Vanilla calls this function whenever an entity who is visible in the turn sequence bar is pushed back in the sequence resulting in calling `entity.onTurnResumed` on the currently active actor prematurely.

Our fix is a bandaid which returns early when calling this function on an entity already in the first slot.

Bug report: https://steamcommunity.com/app/365360/discussions/1/604155219069147354/