fix: populate hireable backgrounds in FirstWorldInit
This fixes issues whereby certain keys are not properly instantiated by other mods before they are accessed in the create functions of the world entities. For example, if a mod adds a completely new party via Dynamic Spawns Framework (DSF) and then tries to use that as the spawn list of a location, it will throw an error if we instantiate the location AfterHooks because DSF registers parties AfterHooks so the key in ::Const.World.Spawn will not be present when for that location when Modular Vanilla instantiates it.

This change also removes the need for us to temporarily instantiate all the different things instantiated in world_state.onInit which may be accessed by certain locations during their create functions.