fix(vanilla): set corpse.Faction for all dying actors
Vanilla (as of 1.5.1.7) only sets `corpse.Faction` for a subset of actors i.e. `human`, `skeleton`, `zombie` and the ones dying from `obsidian_dagger.onDamageDealt`. This leads to inconsistencies when mods try to check the faction of a corpse. We set the `corpse.Faction` of all corpses generated by actors to the faction of the dying actor.

Bug report: https://steamcommunity.com/app/365360/discussions/0/652585529159813893/