fix: populate hireable character backgrounds list on demand
The list must now always be fetched by the getter function MV_getHireableCharacterBackgrounds instead of directly accessing it. The first access will now populate it dynamically. This should only be done when settlements can be safely instantiated. This fixes the issue whereby this list was not being populated early enough for origins to be able to access it during `onSpawnAssets` which runs before `FirstWorldInit` bucket of Modular Vanilla.